+++ wincon/pdcclip.c
@@ -70,9 +70,9 @@ int PDC_getclipboard(char **contents, long *length)
     }
 
 #ifdef PDC_WIDE
-    len = wcslen((wchar_t *)handle) * 3;
+    len = (long)wcslen((wchar_t *)handle) * 3;
 #else
-    len = strlen((char *)handle);
+    len = (long)strlen((char *)handle);
 #endif
     *contents = (char *)GlobalAlloc(GMEM_FIXED, len + 1);
 
@@ -83,7 +83,7 @@ int PDC_getclipboard(char **contents, long *length)
     }
 
 #ifdef PDC_WIDE
-    len = PDC_wcstombs((char *)*contents, (wchar_t *)handle, len);
+    len = (long)PDC_wcstombs((char *)*contents, (wchar_t *)handle, len);
 #else
     strcpy((char *)*contents, (char *)handle);
 #endif
